Update README for script, add initial recordings, cleanup nock helper
