+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019180708590-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1100","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODQyNywiZXhwIjoxNTcxNTEyMDI3LCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.JkogHKZ1hPMpCGVzGokKWY0_0-aCTfiJwMavCuIhy27P5ze5f27xrnEIVdh--3GrS-JPQaMPDyAyuNoAgT7XtXaGdUuEEYv19TMU9W5LBcpT21qqCwKTbEgmEiTwpqlR4XU2hVniNVJk0lUNZ47vLS5Hm878awZH8fSVV_PIhZZvS4a29Sc1ioFMwrKX8hPKeWqFtiZ1DD7OQrbsSyhfDH4zzjs3VA43BSQM5bjBQNu8ugwTM4Z6j66EW7-wGUntO5QBniRN4OyZRHXpdn7tYzRS3PVpcxNV-VhRO_vrPdtEaY7qKhSQa2kMpsxtuBU4lphaJpxlwHwW_n0UE3mtAQ","content-type":"application/json; charset=utf-8","contextid":"tcid=5090279026567115162, server=BL2PEPF00000EBA","ms-cv":"0sCZyyvuHE2CyYmDoalo1g.1","x-forwarded-for":"13.68.16.246","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"name":"composeExtension/fetchTask","type":"invoke","timestamp":"2019-10-19T18:07:07.470Z","localTimestamp":"2019-10-19T18:07:07.470Z","id":"f:5090279026567115162","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","name":"Dave Taniguchi","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"entities":[{"locale":"en-US","country":"US","platform":"Windows","type":"clientInfo"}],"channelData":{"channel":{"id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"},"source":{"name":"compose"}},"value":{"commandId":"createWithPreview","commandContext":"","context":{"theme":"default"}},"locale":"en-US"}}
