+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019180712488-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"2144","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODQzMSwiZXhwIjoxNTcxNTEyMDMxLCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.JFmtUh1gM23djtEcIr84GV5Oea57zM-XZhrRKZRWtjjLPqnFY9z27YvhQOAX5LS9xTWY0Wlds5MD64XXvTRP_uDU4yIPZV-xhy0KgfNE3QtdULWQFmq2AlTPY1aVnkCVpKr2ckM_AdQOF0NxL91zLveBvWz64MTDkrk1SJI4-NWRdiim18a3txcReLL8zswRxufYrb0N4RIBSPL8rzIE5mrlaB_A3Kk03ufZM0-cBUzV3Bx_1WfxK4eZd42Ar52T8uEL56xrCA67QdHIqJAS2-eaDrSR_oxuIITda3VLlJuA-_v6O6NYOvvfgNXl4nWumuTNb40tWDwA6ooMgfFTSQ","content-type":"application/json; charset=utf-8","contextid":"tcid=838328757784002706, server=BL2PEPF000001C5","ms-cv":"Bs9n3rxkLUCYTfzpArm8fA.1","x-forwarded-for":"13.68.16.246","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"name":"composeExtension/submitAction","type":"invoke","timestamp":"2019-10-19T18:07:11.789Z","localTimestamp":"2019-10-19T18:07:11.789Z","id":"f:838328757784002706","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","name":"Dave Taniguchi","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"entities":[{"locale":"en-US","country":"US","platform":"Windows","type":"clientInfo"}],"channelData":{"channel":{"id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"},"source":{"name":"compose"}},"value":{"commandId":"createWithPreview","commandContext":"","botMessagePreviewAction":"send","botActivityPreview":[{"type":"message/card","attachments":[{"content":{"type":"AdaptiveCard","body":[{"color":null,"horizontalAlignment":null,"isSubtle":false,"maxLines":0,"size":null,"text":"Adaptive Card from Task Module","weight":"bolder","wrap":false,"separator":false,"type":"TextBlock"},{"color":null,"horizontalAlignment":null,"isSubtle":false,"maxLines":0,"size":null,"text":"","weight":null,"wrap":false,"id":"Question","separator":false,"type":"TextBlock"},{"isMultiline":false,"maxLength":0,"placeholder":"Answer here...","style":null,"isRequired":false,"id":"Answer","separator":false,"type":"Input.Text"},{"choices":[{"title":"","value":""},{"title":"","value":""},{"title":"","value":""}],"isMultiSelect":true,"style":"expanded","isRequired":false,"id":"Choices","separator":false,"type":"Input.ChoiceSet"}],"actions":[{"data":{"submitLocation":"messagingExtensionSubmit"},"title":"Submit","type":"Action.Submit"}],"version":"1.0"},"contentType":"application/vnd.microsoft.card.adaptive"}]}],"context":{"theme":"default"}},"locale":"en-US"}}
