+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019180714237-reply-integrationBot.json
@@ -1,0 +1,1 @@
+{"scope":"https://smba.trafficmanager.net:443","method":"POST","path":"/amer/v3/conversations","body":{"isGroup":true,"activity":{"type":"message","attachments":[{"contentType":"application/vnd.microsoft.card.adaptive","content":{"actions":[{"type":"Action.Submit","title":"Submit","data":{"submitLocation":"messagingExtensionSubmit"}}],"body":[{"text":"Adaptive Card from Task Module","type":"TextBlock","weight":"bolder"},{"text":"","type":"TextBlock","id":"Question"},{"id":"Answer","placeholder":"Answer here...","type":"Input.Text"},{"choices":[{"title":"","value":""},{"title":"","value":""},{"title":"","value":""}],"id":"Choices","isMultiSelect":true,"style":"expanded","type":"Input.ChoiceSet"}],"type":"AdaptiveCard","version":"1.0"}}]},"channelData":{"channel":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"}}},"status":201,"response":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype;messageid=1571508433697","activityId":"1:1APaiLHstZH0JuPjaBMYqSThrMjaK3ct_ZEiQed5rsdk"},"rawHeaders":["Content-Length","143","Content-Type","application/json; charset=utf-8","Server","Microsoft-HTTPAPI/2.0","Date","Sat, 19 Oct 2019 18:07:13 GMT","Connection","close"],"reqheaders":{"accept":"application/json, text/plain, */*","authorization":"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6ImFQY3R3X29kdlJPb0VOZzNWb09sSWgydGlFcyIsImtpZCI6ImFQY3R3X29kdlJPb0VOZzNWb09sSWgydGlFcyJ9.eyJhdWQiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiaXNzIjoiaHR0cHM6Ly9zdHMud2luZG93cy5uZXQvZDZkNDk0MjAtZjM5Yi00ZGY3LWExZGMtZDU5YTkzNTg3MWRiLyIsImlhdCI6MTU3MTUwODEzMiwibmJmIjoxNTcxNTA4MTMyLCJleHAiOjE1NzE1MTIwMzIsImFpbyI6IjQyVmdZS2o1ZERhL1I1VTV5ZDFiTlNiLzFPemxBQT09IiwiYXBwaWQiOiJlYWM1ZTVjNi1hYjE0LTRmNTAtODBjYi00OTIxM2ZlNTFiMWEiLCJhcHBpZGFjciI6IjEiLCJpZHAiOiJodHRwczovL3N0cy53aW5kb3dzLm5ldC9kNmQ0OTQyMC1mMzliLTRkZjctYTFkYy1kNTlhOTM1ODcxZGIvIiwidGlkIjoiZDZkNDk0MjAtZjM5Yi00ZGY3LWExZGMtZDU5YTkzNTg3MWRiIiwidXRpIjoibVFyMzcySjNZVXkzY1VVV2xjY1JBUSIsInZlciI6IjEuMCJ9.odFal4nVzWvfmVi9IoeWN1APMTFzYPhOtrAZ4JAlTN0T_kFERY9YTvmTmZTPx3imWOQgP_j_LqagsktKovSfebYQe3Ufy8U_M7-93XjUek9gRgtetuMt1Ma9CdsPAc5VIlesXcKkDE9xZv0JUUzYEYyUo4hdgoaK41Y0F__38sMNTh0c90yBcyZ_h1zLmwd51N2xHf_syCQ8LX60hiRer2Dc8JOIGOXl9U-JOEw2vodXVfhrkkXJflcbSafxrT6AUzOTllJW6qi7Gnbfn8IeCFZi_rYh2__U4WhAAhlbOr3wTHgzlyjQjrKMr-RVelZip6bp6LEU4tJ4p2Q0_mZcXA","cookie":"","content-type":"application/json; charset=utf-8","content-length":729,"host":"smba.trafficmanager.net"}}
