+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019180810156-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1044","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODQ4OSwiZXhwIjoxNTcxNTEyMDg5LCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.L2miZinOR0qMzIFmeSo8_a7rZP2gji_UfKC2c_4ZsaMPDIzKPxzTwbJp8cveKg0KHBa8Bik-s9HH00Dyw2qb5Skd4Us-Xzj9E3I06FXxLL_93tIA5TJ1nOhhzs9bPhsHGf76LAGbJMPyunNdmaDLmmNkZd14uOaeiuPzvmhCDBWavKklHzWhNC9hyWB4xsj8xosyb1kX375oIsVCMxH6JXc9TE0eE-t6_uCcZKwQvHMjtnuivnuQ5BXvaSg2pwIQKORlO9cwtQy9xwx5LdGkj3iwl1LxFANoGMtangCEVEPo8JvFgMfPuCSgNVOmQ10Grl7UhWGez7T3sA0eR9uk8g","content-type":"application/json; charset=utf-8","contextid":"tcid=209955408936074601, server=BL2PEPF00000EC7","ms-cv":"zb07VICRPkKsou561GhRKw.1","x-forwarded-for":"13.68.16.246","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"name":"composeExtension/queryLink","type":"invoke","timestamp":"2019-10-19T18:08:09.576Z","localTimestamp":"2019-10-19T18:08:09.576Z","id":"f:209955408936074601","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","name":"Dave Taniguchi","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"entities":[{"locale":"en-US","country":"US","platform":"Windows","type":"clientInfo"}],"channelData":{"channel":{"id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"},"source":{"name":"compose"}},"value":{"url":"https://foo.com/"},"locale":"en-US"}}
