+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019180812506-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1604","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODQ5MSwiZXhwIjoxNTcxNTEyMDkxLCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.HEZwJkNdbajWpKg_oRovTDlTQMFIWQv9TWtuTPljCOzJh4iQXkvZCooAzpNZEfAlyJYH0V82BZ9YACRwJyn-MmhD9emNPTNShYGlO5f-v94MRSyZQ_bT9RL9d6F9pYP3l-UL56HjNd_Z3TsF5viQZ6SUQu1R5rdyUxa_02xRNgkVegbRh8EBxjJwNmWdTiTA1CPNE3KX-5o2YKph3w99Lo0Zl1aw8glr-u4WGg9Vd_v2-jQg6xV-hRzWtATDGMAkWLRG6PDncQd6NKVyX1P55n7rggConzKzhke1_7L5XzL4HSm0bjmZUK9xydvbeSOMoorLOwQDq6rlu3gWo4vR3g","content-type":"application/json; charset=utf-8","contextid":"tcid=2023063120161132813, server=BL2PEPF00000EB5","ms-cv":"pDDXRNUdsUCzWcgPSccSrQ.1","x-forwarded-for":"13.68.16.246","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"text":"<at>My Combined</at> https://foo.com \n","textFormat":"plain","attachments":[{"contentType":"text/html","content":"<div><div><span itemscope=\"\" itemtype=\"http://schema.skype.com/Mention\" itemid=\"0\">My Combined</span>&nbsp;<a href=\"https://foo.com\" rel=\"noreferrer noopener\" target=\"_blank\" title=\"https://foo.com\">https://foo.com</a>&nbsp;</div>\n</div>"}],"type":"message","timestamp":"2019-10-19T18:08:11.810Z","localTimestamp":"2019-10-19T18:08:11.810Z","id":"1571508491752","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","name":"Dave Taniguchi","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype;messageid=1571508491752"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"entities":[{"mentioned":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"My Combined"},"text":"<at>My Combined</at>","type":"mention"},{"locale":"en-US","country":"US","platform":"Windows","type":"clientInfo"}],"channelData":{"teamsChannelId":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype","teamsTeamId":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype","channel":{"id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"}},"locale":"en-US"}}
