+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181135720-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"831","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODY5NCwiZXhwIjoxNTcxNTEyMjk0LCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.W0yly_fGBVIR_9FC3jbsfw66B-sWT9_gqEDP3Je_T6JKndjyVPVP6gC81bIwzNLlmcEQmpR9WHYJjCXDqfzQZjUqRbuQN_62Ln_A5ypMAq2_aZz8jksRkDxu0Wl0JI2H-h_tetnZRSaVOl4TixbEVs6B9YqadKqr1Wx8svKQzV-759-DbcbnVWxHfEKw2gP0R3qFZXFBj7OSKFODPSKT6PM_rROA9oqPVWzXP8qaA889Qbaci3lhcZ4S68_ZL-MQ-bH-WxeBDFCmX5gvoSTp7ju65KWEcBrw5iFDxmLZ9P7uBfAA45QktnQ7_D0rMMVxpRJ-H0c8igI5zXzv27F1BA","content-type":"application/json; charset=utf-8","contextid":"tcid=3382555749578417248, server=BY3PEPF000034DE","ms-cv":"Mwa3f4CHF0y+kC5VOKcMJw.1","x-forwarded-for":"13.91.101.57","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"type":"conversationUpdate","timestamp":"2019-10-19T18:11:34.905Z","id":"f:090e2569-a0c4-727d-e710-bd320c1ff2a4","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"channelData":{"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype","name":"DaveTeams TEST","aadGroupId":"2c02739d-4041-4ef2-b8b8-75508bc04808"},"eventType":"teamRenamed","tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"}}}}
