+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181153901-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"848","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODcxMywiZXhwIjoxNTcxNTEyMzEzLCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.TgO0VnAPtygCFNRDpIfmMOPmLYMjC2nsopgvvIuUCep_sXm1BWjUEiFSk6DHAlw1BcetB9_NpNN27yrR2-Tu1ov1YC_EFzlr9LpXEvK-MU3ArKLtoU5_ey09i0HU3MH33VBArexJzcxghH9fIQrntVtoQk3jSHdMjd4skbeQxpzXHLQykN84gEUDvh7q1Qgt7NnBKQNYqqPx0UPNyICuh_kKJBiaeP5jUjGfaHT4_zEeILDkbmIpe5gnsFhSaOJzAiCWh-y7Jb_nolYTAlBzD1tJvS6FJ4UWkA71aRrnVqZ9QszauBKh5j9JbtnAMC4GO2EoergV05NE_PifTS1Osw","content-type":"application/json; charset=utf-8","contextid":"tcid=4290543649104576324, server=BY3PEPF000034E4","ms-cv":"t1cM+i2E1Eu0x+fDE4snSg.1","x-forwarded-for":"13.91.101.57","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"type":"conversationUpdate","timestamp":"2019-10-19T18:11:53.321Z","id":"f:80c7fbd4-f724-a35b-29cc-195ab0c666b8","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"channelData":{"channel":{"id":"19:9d3c71104f8c49128368b23a5eaeb86d@thread.skype","name":"TEST Channel"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"eventType":"channelRenamed","tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"}}}}
