+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181154245-reply-integrationBot.json
@@ -1,0 +1,1 @@
+{"scope":"https://smba.trafficmanager.net:443","method":"POST","path":"/amer/v3/conversations/19%3A097e6717fd7245bdbeba6baa13840db8%40thread.skype/activities/f%3A80c7fbd4-f724-a35b-29cc-195ab0c666b8","body":{"type":"message","serviceUrl":"https://smba.trafficmanager.net/amer/","channelId":"msteams","from":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"conversation":{"isGroup":true,"conversationType":"channel","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47"},"recipient":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"attachmentLayout":"list","inputHint":"acceptingInput","attachments":[{"contentType":"application/vnd.microsoft.card.hero","content":{"title":"Channel Renamed","text":"TEST Channel is the new Channel name"}}],"replyToId":"f:80c7fbd4-f724-a35b-29cc-195ab0c666b8"},"status":201,"response":{"id":"1:1gPaleHBoA4Yc8pl1y3lWl9cX6tYArzsej1YoTO6ADNw"},"rawHeaders":["Content-Length","55","Content-Type","application/json; charset=utf-8","Server","Microsoft-HTTPAPI/2.0","Date","Sat, 19 Oct 2019 18:11:52 GMT","Connection","close"],"reqheaders":{"accept":"application/json, text/plain, */*","authorization":"Bearer eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIsIng1dCI6ImFQY3R3X29kdlJPb0VOZzNWb09sSWgydGlFcyIsImtpZCI6ImFQY3R3X29kdlJPb0VOZzNWb09sSWgydGlFcyJ9.eyJhdWQiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiaXNzIjoiaHR0cHM6Ly9zdHMud2luZG93cy5uZXQvZDZkNDk0MjAtZjM5Yi00ZGY3LWExZGMtZDU5YTkzNTg3MWRiLyIsImlhdCI6MTU3MTUwODEzMiwibmJmIjoxNTcxNTA4MTMyLCJleHAiOjE1NzE1MTIwMzIsImFpbyI6IjQyVmdZS2o1ZERhL1I1VTV5ZDFiTlNiLzFPemxBQT09IiwiYXBwaWQiOiJlYWM1ZTVjNi1hYjE0LTRmNTAtODBjYi00OTIxM2ZlNTFiMWEiLCJhcHBpZGFjciI6IjEiLCJpZHAiOiJodHRwczovL3N0cy53aW5kb3dzLm5ldC9kNmQ0OTQyMC1mMzliLTRkZjctYTFkYy1kNTlhOTM1ODcxZGIvIiwidGlkIjoiZDZkNDk0MjAtZjM5Yi00ZGY3LWExZGMtZDU5YTkzNTg3MWRiIiwidXRpIjoibVFyMzcySjNZVXkzY1VVV2xjY1JBUSIsInZlciI6IjEuMCJ9.odFal4nVzWvfmVi9IoeWN1APMTFzYPhOtrAZ4JAlTN0T_kFERY9YTvmTmZTPx3imWOQgP_j_LqagsktKovSfebYQe3Ufy8U_M7-93XjUek9gRgtetuMt1Ma9CdsPAc5VIlesXcKkDE9xZv0JUUzYEYyUo4hdgoaK41Y0F__38sMNTh0c90yBcyZ_h1zLmwd51N2xHf_syCQ8LX60hiRer2Dc8JOIGOXl9U-JOEw2vodXVfhrkkXJflcbSafxrT6AUzOTllJW6qi7Gnbfn8IeCFZi_rYh2__U4WhAAhlbOr3wTHgzlyjQjrKMr-RVelZip6bp6LEU4tJ4p2Q0_mZcXA","cookie":"","content-type":"application/json; charset=utf-8","content-length":759,"host":"smba.trafficmanager.net"}}
