+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181214767-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1026","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODczMywiZXhwIjoxNTcxNTEyMzMzLCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.MJeZZckncX7iGK21YkvEYktfNNdgvMMZc_JHc2HW3u23_jPh2QNOJNWzUOq0T0SgtL1JSFBEl86LFIjM1cgvTLxNcOYOqvvs_myM0SrGCF3i9f7d0EZDHGuwkZFW_k0lDZarboiNhER37l9STc3j8IorBdPBcvdTRjECQuochH15HezSCyEvGSo41ksxcdf1l1YR7W5czFLBkaDyLZTbBBXPuDNrnLE6j6YyyQ0S7F2ZQOR8EdMFtEuKUGblUFhOUNeWaoH21jzOvP7l4mlwsSXa2y2tfWjFiWauD4q79GcwE0l9fVE2Ed_eKGSsyejEoj-j3VP7arU2Z0Gi_9qdUw","content-type":"application/json; charset=utf-8","contextid":"tcid=9204825831208946116, server=BY3PEPF000034EB","ms-cv":"4vWa5CKptEC9ZCOS7p/1Tg.1","x-forwarded-for":"13.91.101.57","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"membersAdded":[{"id":"29:1OeqZAirv1xrSRIofsdOhqfIoSEZImcZoJELvXmhzoyn811wFf3whJ557yiWNHLw5VKK3DvhMMFh_Gpdo74o_XIy16bH3DjLUoxphFX5Pi_A","aadObjectId":"e830a11b-50fe-465a-810c-5c308099ca2b"}],"type":"conversationUpdate","timestamp":"2019-10-19T18:12:13.945Z","id":"f:2ad40fba-2b9c-9a8c-d48b-27f72dcc0f71","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"channelData":{"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype","name":"DaveTeams TEST","aadGroupId":"2c02739d-4041-4ef2-b8b8-75508bc04808"},"eventType":"teamMemberAdded","tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"}}}}
