+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181218849-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1030","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODczOCwiZXhwIjoxNTcxNTEyMzM4LCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.O9_v1tATLphA9aNMuHme9MVTmFjAPhDPHQmGEfLJIXNrNnb3dZ6EuY8XO7luTAZzxUCXTxwYxSAn8J348ATpn4mcC_sm0_bStUoVi8PWD9gnJoask4Ok86JDgn7g3JyZW5umP6HaMD1csX8C-5xNpse9htR64kJQd1lEm-EiHNRnxk9_5Rr_8AqRyxK8ao5O45u6hp2vIPyRPMftlvqfmlt1iWBH5jqkUrVyEwzpmcK5TT9PQi7wrf1jE7qyaXvy9WkHKdI4MeJrN_qgINklegYmxZWJQLxSUG3cDUBKZUhxS-WINcUQnuh7jBkmXPbigb_agOj6mcShpp_NgqoLvw","content-type":"application/json; charset=utf-8","contextid":"tcid=519156183249225851, server=BY3PEPF000034F2","ms-cv":"LGBRuFDegkinWHqPjbKVcQ.1","x-forwarded-for":"13.91.101.57","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"membersRemoved":[{"id":"29:1OeqZAirv1xrSRIofsdOhqfIoSEZImcZoJELvXmhzoyn811wFf3whJ557yiWNHLw5VKK3DvhMMFh_Gpdo74o_XIy16bH3DjLUoxphFX5Pi_A","aadObjectId":"e830a11b-50fe-465a-810c-5c308099ca2b"}],"type":"conversationUpdate","timestamp":"2019-10-19T18:12:18.271Z","id":"f:09be3aa0-527f-c36b-2690-02d30b01792f","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"channelData":{"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype","name":"DaveTeams TEST","aadGroupId":"2c02739d-4041-4ef2-b8b8-75508bc04808"},"eventType":"teamMemberRemoved","tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"}}}}
