+++ libraries/botbuilder/tests/teams/integrationBot/recordings/20191019181248209-request-integrationBot.json
@@ -1,0 +1,1 @@
+{"type":"request","url":"/api/messages","method":"POST","headers":{"host":"localhost:3978","user-agent":"Microsoft-SkypeBotApi (Microsoft-BotFramework/3.0)","content-length":"1130","authorization":"Bearer eyJhbGciOiJSUzI1NiIsImtpZCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSIsInR5cCI6IkpXVCIsIng1dCI6IktwSVdSVWxnZmlObGQxRFR4WkFoZTRpTm1rQSJ9.eyJzZXJ2aWNldXJsIjoiaHR0cHM6Ly9zbWJhLnRyYWZmaWNtYW5hZ2VyLm5ldC9hbWVyLyIsIm5iZiI6MTU3MTUwODc2NywiZXhwIjoxNTcxNTEyMzY3LCJpc3MiOiJodHRwczovL2FwaS5ib3RmcmFtZXdvcmsuY29tIiwiYXVkIjoiZWFjNWU1YzYtYWIxNC00ZjUwLTgwY2ItNDkyMTNmZTUxYjFhIn0.k0M6m2TyQ2fEjXp4EZL4Gi8ajyiY1KFXdS4uh_xI1AYd0px-NKA8qD1SGbpEUXEnqUBVQuPao4G0b73h97entwXVd23EbEs1P9cLRbBxKwuY61nk1ij5Bqb1PF6DvbQjIxqNjUBegqgAYMxnXUnQId5RO7bHFuNv3jCc7xZBoG8x559L5rn0WFLFdZqGRQxuW8C3bjO89U7X_BMt3CLS7L0A68jGiGcnvYo1PyrWmYs6s8rncIrLXUaJTS2LaCeLNu96nsLEflRmRfqHZe-noxZXAp7VdZzvI4VUXY1hiWQKV3w7QGz4x0wVk0MzRX-VWun7V-WPSGfKRUCs4KYqPg","content-type":"application/json; charset=utf-8","contextid":"tcid=5435035626002932798, server=BL2PEPF000001DA","ms-cv":"y9OasW4b3kGI18dX85cjgQ.1","x-forwarded-for":"13.68.16.246","x-forwarded-proto":"https","x-original-host":"e00cd789.ngrok.io"},"body":{"name":"composeExtension/query","type":"invoke","timestamp":"2019-10-19T18:12:47.526Z","localTimestamp":"2019-10-19T18:12:47.526Z","id":"f:5435035626002932798","channelId":"msteams","serviceUrl":"https://smba.trafficmanager.net/amer/","from":{"id":"29:1CFIxRJBLjT_io3iNsDZxMk_4YPWQmr3DZUM1I6ihf5L2ocCDqg9S1vkHK6kRdGfK1P_2LFE2nQqFIvdu-GBOZA","name":"Dave Taniguchi","aadObjectId":"6b703ce3-6d14-4fe6-b90c-dc5f03657ae4"},"conversation":{"isGroup":true,"conversationType":"channel","tenantId":"72f988bf-86f1-41af-91ab-2d7cd011db47","id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"recipient":{"id":"28:eac5e5c6-ab14-4f50-80cb-49213fe51b1a","name":"teams2"},"entities":[{"locale":"en-US","country":"US","platform":"Windows","type":"clientInfo"}],"channelData":{"channel":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"team":{"id":"19:097e6717fd7245bdbeba6baa13840db8@thread.skype"},"tenant":{"id":"72f988bf-86f1-41af-91ab-2d7cd011db47"},"source":{"name":"compose"}},"value":{"commandId":"searchQuery","parameters":[{"name":"searchQuery","value":"ex"}],"queryOptions":{"skip":0,"count":25}},"locale":"en-US"}}
